+++ packages/reaxes-react/src/hooks/useReaxable.ts
@@ -1,0 +1,6 @@
+
+export const useReaxable = <S extends object>( state : S ) => useRef( (createReaxable as CreateReaxable)( state ) ).current;
+
+import { createReaxable , type CreateReaxable} from 'reaxes';
+import type {NestedMutate,NestedSetState,ResurivePartial} from 'reaxes';
+import { useRef } from 'react';